requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
alembic
bcrypt
pytest
pyjwt
typer
pytest-asyncio
pytest-cov